Initial support for clicking on hyperlinks
Clicking on a hyperlink will immediately execute it (if it's a game command), or
open it in the external application ("http", "mailto", etc.)  This happens at
the moment the mouse button is pressed.  This is not nice; it needs to happen
when the button is released.  Will work on that later.  Also, for now there's no
visual feedback at all when hovering over and when clicking.