Use SMPEG library to decode mp3 sounds
SDL_Mixer can only play MP3s in the "music" channel, which doesn't allow more
than one sound to be played at the same time.  It can't play MP3s in the regular
mixed channels because it can't decode them due to it's brain damaged API.  We
use SMPEG to decode MP3s and then play them in the regular mixer channels.

This means we decode the *whole* MP3 into memory.  That's not nice, but for now
we have to live with it.