Rewrite banner support code
Banners are now sized and positioned manually instead of relying on QLayouts.
As a result, banners now work as intended; the window is freely resizable when
showing a banner with a 100% height or width, and there are no more weird main
window resizes that seemed to happen on their own.

Among other things, this means that 1893 now runs correctly.