Don't show()/hide() the about box at creation time
This was needed because the viewport dimensions of a scroll area only get
updated when show() is called.  This resulted in the about box flashing in
and out every time one was created.  Now we don't use the viewport anymore
to report the about box size to the formatter but use the about box
dimensions directly since it never has scrollbars.