Always #undef main if it's defined
On some platforms, SDL redefines main in order to provide a platform-specific
main() implementation.  However, Qt handles this too, so things can get weird.
We need to make sure main is not redefined so that Qt can find our own
implementation and SDL will not try to do platform-specific initialization work
(like launching the Cocoa event-loop or setting up the application menu on
OS X, or redirecting stdout and stderr to text files on Windows), which would
break things.