Reduce amount of widget repaints
Don't change the scrollbar position and neither advance the event loop in
fmt_adjust_vscroll().  The immediate text scrolling after input is kind of
nice, but it results in way too much unneeded widget repaints and a glitch
where text is flashing at the top of the main game window.

To compesate for this change, advance the event loop after a newline in input,
so that we scroll down by a line immediately.