Add OpenSSL version detection in appimage qt hook
The AppImage we ship is built on Ubuntu 16.04 to ensure good
compatibility with older Linux distros. But 16.04 ships OpenSSL 1.0 in
its repos, which is very outdated. More recent distros are using 1.1,
which is incompatible with 1.0. So building against 1.0 breaks the QTads
online update check on any modern distro.

Bundling OpenSSL is actually not a good idea. The host's version should
be used, not only for security reasons but also for compatibility
because certificate locations on the filesystem are different between
distros.

We solve this issue by bundling OpenSSL 1.1 in the AppImage but don't
actually use it if the host already provides a 1.1.x version. We only
use the bundled version if the host has an incompatible version (like
1.0 or 3.0.)

This is still not optimal, but there's not much else we can do about it
since it's not possible to build Qt in a such a way that would make it
support both 1.0 and 1.1 at the same time.

As OpenSSL 1.0 is very outdated by now, we bundle 1.1 in the AppImage.
However, this means th